requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg==3.2.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary>=2.9
pyotp>=2.6.0
Pillow>=9.0.0
python-dotenv
social-auth-app-django
django-otp
qrcode[pil]
qrcode
pyotp
django-otp
djangorestframework-simplejwt

